#58 Round scale parameter to int value to prevent misalignment

+++ PatternPaint/patterneditor.cpp
@@ -86,7 +86,8 @@ bool PatternEditor::init(QImage newPattern, bool scaled) {
 
 void PatternEditor::updateGridSize() {
     // Base the widget size on the window height
-    float scale = float(size().height() - 1)/pattern.height();
+    // cast float to int to save rounded scale
+    float scale = static_cast<int>(float(size().height() - 1)/pattern.height());
 
     // Use a square aspect to display the grid
     xScale = scale;
